chore: JS > TS
