+++ static/main.ts
@@ -1,5 +1,15 @@
+import { ManifestDecoder } from "./scripts/decode-manifest";
+import { ManifestRenderer } from "./scripts/render-manifest";
+
 export async function mainModule() {
-  console.log(`Hello from mainModule`);
+  try {
+    const decoder = new ManifestDecoder(window.location.search.substring(1));
+    const decodedManifest = await decoder.decodeManifest();
+    const renderer = new ManifestRenderer(decodedManifest);
+    renderer.renderManifest();
+  } catch (error) {
+    console.error(error);
+  }
 }
 
 mainModule()
