+++ static/manifest-gui.css
@@ -1,4 +1,3 @@
-
 html{background-image: radial-gradient(#000000,#101010);background-color: #101010;/* height: 100vh; */}
 body{margin:0;color:#fff;font-family:"Proxima Nova",sans-serif;background-image:url(./grid-25.png);height: calc(100vh - 24px);}
 controls button{background:#101010}
@@ -49,3 +48,22 @@ button#add:active{background-color: #00ff0020;}
 button#add:active::before{content:"+++"}
 
 #manifest-gui pre{margin: 0 0 16px;}
+
+
+
+/* body {
+  font-family: "Proxima Nova", "Ubiquity Nova", sans-serif;
+  background-color: #06061aff;
+  color: #fff;
+  background-image: url("data:image/png;base64,iVBORw0KGgoAAAANSUhEUgAAABgAAAAYCAYAAADgdz34AAABYWlDQ1BrQ0dDb2xvclNwYWNlRGlzcGxheVAzAAAokWNgYFJJLCjIYWFgYMjNKykKcndSiIiMUmB/yMAOhLwMYgwKicnFBY4BAT5AJQwwGhV8u8bACKIv64LMOiU1tUm1XsDXYqbw1YuvRJsw1aMArpTU4mQg/QeIU5MLikoYGBhTgGzl8pICELsDyBYpAjoKyJ4DYqdD2BtA7CQI+whYTUiQM5B9A8hWSM5IBJrB+API1klCEk9HYkPtBQFul8zigpzESoUAYwKuJQOUpFaUgGjn/ILKosz0jBIFR2AopSp45iXr6SgYGRiaMzCAwhyi+nMgOCwZxc4gxJrvMzDY7v////9uhJjXfgaGjUCdXDsRYhoWDAyC3AwMJ3YWJBYlgoWYgZgpLY2B4dNyBgbeSAYG4QtAPdHFacZGYHlGHicGBtZ7//9/VmNgYJ/MwPB3wv//vxf9//93MVDzHQaGA3kAFSFl7jXH0fsAAAA4ZVhJZk1NACoAAAAIAAGHaQAEAAAAAQAAABoAAAAAAAKgAgAEAAAAAQAAABigAwAEAAAAAQAAABgAAAAAwf1XlwAAACNJREFUSA3t0IEAAAAMBKFHm7/UTaQQWnXDgAEDBgwYMGDgAXaJAz4RVVHYAAAAAElFTkSuQmCC");
+}
+@font-face {
+  font-family: "Ubiquity Nova";
+  font-style: normal;
+  font-weight: 400;
+  src: url(./fonts/ubiquity-nova-standard.eot);
+  src:
+    url(./fonts/ubiquity-nova-standard.eot#iefix) format("embedded-opentype"),
+    url(./fonts/ubiquity-nova-standard.woff) format("woff"),
+    url(./fonts/ubiquity-nova-standard.ttf) format("truetype");
+} */
